# rimway-backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/send-otp": {
      "post": {
        "operationId": "AuthController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/admin/rides": {
      "get": {
        "operationId": "AdminController_getRides",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "AdminController_createPhoneRide",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhoneRideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/rides/{id}": {
      "get": {
        "operationId": "AdminController_getRideDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/rides/{id}/cancel": {
      "post": {
        "operationId": "AdminController_cancelRide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCancelRideDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/admin/captains/{id}": {
      "get": {
        "operationId": "AdminCaptainsController_getCaptainDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      },
      "patch": {
        "operationId": "AdminCaptainsController_updateCaptain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaptainDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminCaptains"
        ]
      }
    },
...
    "/api/v1/admin/wallet/top-ups": {
      "get": {
        "operationId": "AdminWalletController_getTopUps",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
    "/api/v1/admin/wallet/top-ups/{id}/approve": {
      "post": {
        "operationId": "AdminWalletController_approveTopUp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
    "/api/v1/admin/wallet/top-ups/{id}/reject": {
      "post": {
        "operationId": "AdminWalletController_rejectTopUp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
...
    "/api/v1/admin/wallet/captains/{id}": {
      "get": {
        "operationId": "AdminWalletController_getCaptainWallet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
    "/api/v1/admin/wallet/captains/{id}/transactions": {
      "get": {
        "operationId": "AdminWalletController_getCaptainTransactions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminWallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "RIM WAY API",
    "description": "The RIM WAY Backend API Contract for Passenger and Driver Apps",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "SendOtpDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePhoneRideDto": {
        "type": "object",
        "properties": {
          "pickupLat": {
            "type": "number"
          },
          "pickupLng": {
            "type": "number"
          },
          "dropoffLat": {
            "type": "number"
          },
          "dropoffLng": {
            "type": "number"
          },
          "serviceTypeId": {
            "type": "string"
          },
          "pickupName": {
            "type": "string"
          },
          "dropoffName": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "estimatedFare": {
            "type": "number"
          },
          "distanceKm": {
            "type": "number"
          }
        },
        "required": [
          "pickupLat",
          "pickupLng",
          "dropoffLat",
          "dropoffLng",
          "serviceTypeId",
          "pickupName",
          "dropoffName",
          "customerName",
          "customerPhone"
        ]
      },
      "AdminCancelRideDto": {
        "type": "object",
        "properties": {
          "stateVersion": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "stateVersion"
        ]
      },
...
      "UpdateCaptainDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}